trading_bot/requirements.txt
discord.py==1.7.3
fastapi==0.95.0
uvicorn==0.17.6
python-dotenv==0.19.2
jinja2==3.0.3
requests==2.27.1
